backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
pydantic==1.10.12
sqlalchemy==2.0.49
alembic==1.14.1
asyncpg==0.30.0
gunicorn==23.0.0
aiosqlite==0.21.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
numpy==2.4.4
scikit-learn==1.8.0
bcrypt==5.0.0
qrcode==8.2
pillow==11.3.0
